Refactor unsplash logic. Fix input[type='text'] being removed from minified version affecting some styling
