+++ options.html
@@ -36,7 +36,7 @@
                     <div id="custom-city-container" style="display: none;">
                         <label for="custom-city">Custom City</label>
                         <div id="city-search-container">
-                            <input type="text" id="custom-city" name="custom-city" placeholder="Search for a city..." autocomplete="off">
+                            <input class="text-input" type="text" id="custom-city" name="custom-city" placeholder="Search for a city..." autocomplete="off">
                             <div id="city-suggestions"></div>
                         </div>
                     </div>
@@ -110,7 +110,7 @@
                     <p style="font-size: 0.9em; opacity: 0.8; margin-top: -0.5em; margin-bottom: 1em;">This will fetch a new high-quality background image. It will override the custom background image if one is set.</p>
                     <div class="setting-row">
                         <label for="unsplash-api-key">Unsplash API Key (Required)</label>
-                        <input type="text" id="unsplash-api-key" placeholder="Enter your Unsplash Access Key">
+                        <input class="text-input" type="text" id="unsplash-api-key" placeholder="Enter your Unsplash Access Key">
                         <p style="font-size: 0.8em; opacity: 0.7; margin-top: 0.3em;">An API key is required to fetch images from Unsplash. Get a free key from <a href="https://unsplash.com/developers" target="_blank" rel="noopener">Unsplash Developers</a>.</p>
                     </div>
                     <div class="tooltip-wrapper" data-tooltip="Requires an Unsplash API Key">
